<update> update player operations item for inventory

+++ GameLogic/src/main/java/nro/model/item/ItemTemplate.java
@@ -3,7 +3,7 @@ package nro.model.item;
 import java.util.List;
 
 public record ItemTemplate(short id, byte type, byte gender, String name, String description, byte level, short iconID,
-                           short part, boolean isUpToUp, int strRequire, int head, int body, int leg, List<ItemOption> options) {
+                           short part, boolean isUpToUp, int strRequire, short head, short body, short leg, List<ItemOption> options) {
 
     public record ArrHead2Frames(int id, List<Integer> frames) {
     }
